changes up to now

+++ code/delay.h
@@ -3,8 +3,8 @@
 
 #include <util/delay.h>
 
-void delay_ms (int ms);
-void delay_s (int s);
+void delay_ms (int16_t ms);
+void delay_s (int16_t s);
 void delay_us(uint16_t us);
 
 #endif /* DELAY_H */
